DH-367 | Sort/Filter 아이콘 추가 및 GUI 수정

+++ dailyHOTEL/app/src/main/java/com/twoheart/dailyhotel/model/GourmetFilter.java
@@ -27,8 +27,8 @@ public class GourmetFilter implements Parcelable
 
     public GourmetFilter(JSONObject jsonObject) throws JSONException
     {
-        long openTimeInMillis = jsonObject.getInt("startEatingTime");
-        long closeTimeInMillis = jsonObject.getInt("endEatingTime");
+        long openTimeInMillis = jsonObject.getLong("startEatingTime");
+        long closeTimeInMillis = jsonObject.getLong("endEatingTime");
 
         timeFlag = getTimeFlag(openTimeInMillis) | getTimeFlag(closeTimeInMillis);
